Add some Chef stuff
To help manage the server. Adds a user and sets up Docker.

At the moment I run the service by running it as a container, although I
can't manually deploy it yet because the SLACK_API_TOKEN is sensitive
data. Once that's been added a Docker secret then we can deploy using
Chef to fetch the latest image.