Reorganise structure, add hooks
This commit fleshes out the structure of the bot.

It adds some hooks to comment when a user enters or leaves a channel.

I've chosen to separate the hooks and commands so it's easy to copy and
add new ones without having huge files of different commands. The
commands are mostly going to be distinct, and so each file for each type
of command felt right.

It should only be neccessary to add a new command in the directory,
rebuild the image, push and redeploy.