Add weather command
This command scans for the phrase "weather in <location>" and will tell you roughly
what the weather will be like in that location. It uses the Met Office
Datapoint API[1] to fetch the likely location (first doing an exact
match, then trying a fuzzy match) and will return the temperature and
chance of rain in a human readable way.

[1] https://www.metoffice.gov.uk/datapoint/support/api-reference